[Other] Optimize codes using list comprehension, implement GraphCompilerBackend for paddle test_compiler. (#327)
* Optimize codes using list comprehension.

* Implement graph compiler backend for paddle.

* Fix the init dtype.

+++ graph_net/paddle/utils.py
@@ -213,7 +213,7 @@ def replay_tensor(info):
         ).to(device)
     else:
         if mean is not None and std is not None:
-            tensor = paddle.empty(shape=shape, dtype=dtype)
+            tensor = paddle.empty(shape=shape, dtype="float32")
             initializer = paddle.nn.initializer.TruncatedNormal(
                 mean=mean, std=std, a=min_val, b=max_val
             )
